Extract story/tests to separate file
