+++ packages/reakit/src/Tabbable/__examples__/index.story.ts
@@ -1,6 +1,7 @@
 import { Tabbable } from "../Tabbable";
 
 export { default as TabbableElements } from "./TabbableElements";
+export { default as TabbableWithFocusVisible } from "./TabbableWithFocusVisible";
 
 export default {
   title: "Tabbable",
